Remove non-functioning Library drop-down
